+++ packages/brick-kit/src/core/Router.ts
@@ -1,5 +1,10 @@
 import { PluginLocation } from "@easyops/brick-types";
-import { loadScript } from "@easyops/brick-utils";
+import {
+  loadScript,
+  getTemplateDepsOfStoryboard,
+  getDllAndDepsOfStoryboard,
+  processStoryboard
+} from "@easyops/brick-utils";
 import {
   LocationContext,
   mountTree,
@@ -11,6 +16,7 @@ import {
 import { getHistory } from "../history";
 import { httpErrorToString } from "../handleHttpError";
 import { isUnauthenticatedError } from "../isUnauthenticatedError";
+import { brickTemplateRegistry } from "./TemplateRegistries";
 
 export class Router {
   private defaultCollapsed = false;
@@ -33,8 +39,9 @@ export class Router {
   private render = async (location: PluginLocation): Promise<void> => {
     const history = getHistory();
     const locationContext = new LocationContext(this.kernel, location);
+    const { bootstrapData } = this.kernel;
     const storyboard = locationContext.matchStoryboard(
-      this.kernel.bootstrapData.storyboards
+      bootstrapData.storyboards
     );
     if (storyboard) {
       // 如果找到匹配的 storyboard，那么加载它的依赖库。
@@ -46,11 +53,26 @@ export class Router {
           )
         );
         await loadScript(
-          this.kernel.bootstrapData.brickPackages.map(bp => bp.filePath)
+          bootstrapData.brickPackages
+            .map(item => item.filePath)
+            .concat(bootstrapData.templatePackages.map(item => item.filePath))
         );
-      } else {
-        await loadScript(storyboard.dll);
-        await loadScript(storyboard.deps);
+      } else if (!storyboard.depsProcessed) {
+        // 先加载模板
+        const templateDeps = getTemplateDepsOfStoryboard(
+          storyboard,
+          bootstrapData.templatePackages
+        );
+        await loadScript(templateDeps);
+        // 加载模板后才能加工得到最终的构件表
+        const result = getDllAndDepsOfStoryboard(
+          processStoryboard(storyboard, brickTemplateRegistry),
+          bootstrapData.brickPackages
+        );
+        await loadScript(result.dll);
+        await loadScript(result.deps);
+        // 每个 storyboard 仅处理一次依赖
+        storyboard.depsProcessed = true;
       }
     }