+++ packages/brick-kit/src/core/TemplateRegistries.spec.ts
@@ -1,0 +1,13 @@
+import {
+  registerBrickTemplate,
+  brickTemplateRegistry
+} from "./TemplateRegistries";
+
+describe("registerBrickTemplate", () => {
+  it("should ignore duplicated templates", () => {
+    registerBrickTemplate("a", jest.fn());
+    expect(brickTemplateRegistry.size).toBe(1);
+    registerBrickTemplate("a", jest.fn());
+    expect(brickTemplateRegistry.size).toBe(1);
+  });
+});