+++ packages/brick-dll/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.13](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@0.7.12...@easyops/brick-dll@0.7.13) (2019-10-14)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [0.7.12](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@0.7.11...@easyops/brick-dll@0.7.12) (2019-10-14)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [0.7.11](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@0.7.10...@easyops/brick-dll@0.7.11) (2019-10-12)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [0.7.10](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@0.7.9...@easyops/brick-dll@0.7.10) (2019-10-12)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [0.7.9](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@0.7.8...@easyops/brick-dll@0.7.9) (2019-10-12)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [0.7.8](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@0.7.7...@easyops/brick-dll@0.7.8) (2019-10-11)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
 ## [0.7.7](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@0.7.6...@easyops/brick-dll@0.7.7) (2019-10-10)
 
 **Note:** Version bump only for package @easyops/brick-dll