+++ packages/brick-icons/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-icons@0.3.0...@easyops/brick-icons@0.4.0) (2019-10-12)
+
+### Features
+
+- **brick-icon:** 添加巡检图标 ([0c0a69f](https://git.easyops.local/anyclouds/next-core/commits/0c0a69f))
+
+# [0.3.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-icons@0.2.1...@easyops/brick-icons@0.3.0) (2019-10-11)
+
+### Features
+
+- **brick-icon:** 添加 launchpad 大小图标 ([5603f69](https://git.easyops.local/anyclouds/next-core/commits/5603f69))
+
 ## [0.2.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-icons@0.2.0...@easyops/brick-icons@0.2.1) (2019-10-10)
 
 **Note:** Version bump only for package @easyops/brick-icons