+++ packages/build-config-factory/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.2](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@0.17.1...@easyops/build-config-factory@0.17.2) (2019-10-14)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+## [0.17.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@0.17.0...@easyops/build-config-factory@0.17.1) (2019-10-14)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+# [0.17.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@0.16.0...@easyops/build-config-factory@0.17.0) (2019-10-12)
+
+### Features
+
+- ensure and generate templates deps ([47c3b19](https://git.easyops.local/anyclouds/next-core/commits/47c3b19))
+
+# [0.16.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@0.15.2...@easyops/build-config-factory@0.16.0) (2019-10-12)
+
+### Features
+
+- post build of templates ([4da32a2](https://git.easyops.local/anyclouds/next-core/commits/4da32a2))
+- revert using bricks as peer dependencies of micro-apps ([fd3e7a1](https://git.easyops.local/anyclouds/next-core/commits/fd3e7a1))
+
+## [0.15.2](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@0.15.1...@easyops/build-config-factory@0.15.2) (2019-10-12)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
 ## [0.15.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@0.15.0...@easyops/build-config-factory@0.15.1) (2019-10-10)
 
 **Note:** Version bump only for package @easyops/build-config-factory